+++ frontend/src/App.js
@@ -3,6 +3,7 @@ import { Routes, Route, Link } from 'react-router-dom';
 import "./App.css"
 import Home from './Home';
 import Temas from './temas';
+import Alumnos from './Alumnos';
 import CrearCurso from "./crearCurso"
 
 function App() {
@@ -11,11 +12,13 @@ function App() {
       <div class="topnav">
         <Link to="/"><a class="navbarBoton">Precio Compra</a></Link>
         <Link to="/temas"><a class="navbarBoton">Cursos</a></Link>
+        <Link to="/alumnos"><a class="navbarBoton">Alumnos</a></Link>
       </div>
       <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/temas/*" element={<Temas />} />
-          <Route path="/crearCurso/:idTema/:idCurso" element={<CrearCurso />} />
+          <Route path="/alumnos" element={<Alumnos />} />
+          <Route path="/crearCurso/:idTema" element={<CrearCurso />} />
       </Routes>
     </body>
   );